requirements.txt
attrs==21.2.0
aws-cdk.assertions==1.136.0
aws-cdk.assets==1.137.0
aws-cdk.aws-accessanalyzer==1.136.0
aws-cdk.aws-acmpca==1.137.0
aws-cdk.aws-apigateway==1.137.0
aws-cdk.aws-applicationautoscaling==1.137.0
aws-cdk.aws-autoscaling==1.137.0
aws-cdk.aws-autoscaling-common==1.137.0
aws-cdk.aws-autoscaling-hooktargets==1.137.0
aws-cdk.aws-certificatemanager==1.137.0
aws-cdk.aws-cloudformation==1.137.0
aws-cdk.aws-cloudfront==1.137.0
aws-cdk.aws-cloudwatch==1.137.0
aws-cdk.aws-codebuild==1.137.0
aws-cdk.aws-codecommit==1.137.0
aws-cdk.aws-codedeploy==1.137.0
aws-cdk.aws-codeguruprofiler==1.137.0
aws-cdk.aws-codepipeline==1.137.0
aws-cdk.aws-codepipeline-actions==1.137.0
aws-cdk.aws-codestarnotifications==1.137.0
aws-cdk.aws-cognito==1.137.0
aws-cdk.aws-ec2==1.137.0
aws-cdk.aws-ecr==1.137.0
aws-cdk.aws-ecr-assets==1.137.0
aws-cdk.aws-ecs==1.137.0
aws-cdk.aws-efs==1.137.0
aws-cdk.aws-elasticloadbalancing==1.137.0
aws-cdk.aws-elasticloadbalancingv2==1.137.0
aws-cdk.aws-events==1.137.0
aws-cdk.aws-events-targets==1.137.0
aws-cdk.aws-globalaccelerator==1.137.0
aws-cdk.aws-iam==1.137.0
aws-cdk.aws-kinesis==1.137.0
aws-cdk.aws-kinesisfirehose==1.137.0
aws-cdk.aws-kms==1.137.0
aws-cdk.aws-lambda==1.137.0
aws-cdk.aws-logs==1.137.0
aws-cdk.aws-route53==1.137.0
aws-cdk.aws-route53-targets==1.137.0
aws-cdk.aws-s3==1.137.0
aws-cdk.aws-s3-assets==1.137.0
aws-cdk.aws-s3-deployment==1.137.0
aws-cdk.aws-sam==1.137.0
aws-cdk.aws-secretsmanager==1.137.0
aws-cdk.aws-servicediscovery==1.137.0
aws-cdk.aws-signer==1.137.0
aws-cdk.aws-sns==1.137.0
aws-cdk.aws-sns-subscriptions==1.137.0
aws-cdk.aws-sqs==1.137.0
aws-cdk.aws-ssm==1.137.0
aws-cdk.aws-stepfunctions==1.137.0
aws-cdk.aws-wafv2==1.136.0
aws-cdk.cloud-assembly-schema==1.137.0
aws-cdk.core==1.137.0
aws-cdk.custom-resources==1.137.0
aws-cdk.cx-api==1.137.0
aws-cdk.lambda-layer-awscli==1.137.0
aws-cdk.pipelines==1.137.0
aws-cdk.region-info==1.137.0
cattrs==1.8.0
Cerberus==1.3.3
certifi==2020.12.5
chardet==4.0.0
cloudcomponents.cdk-lambda-at-edge-pattern==1.48.0
colorama==0.4.4
colorlog==5.0.1
constructs==3.3.161
distlib==0.3.4
docopt==0.6.2
filelock==3.4.2
idna==2.10
Jinja2==2.11.3
jsii==1.47.0
MarkupSafe==1.1.1
numpy==1.21.1
pipreqs==0.4.11
platformdirs==2.4.1
publication==0.0.3
python-dateutil==2.8.2
PyYAML==5.4.1
pyyaml-include==1.2.post2
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
torch==1.9.0
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.13.0
yarg==0.1.9
